ADD: added rudimentary help system ADD: added binary and source packaging

+++ trunk/Ready/utils.cpp
@@ -32,7 +32,7 @@
     {
         struct _timeb timebuffer;
         _ftime_s( &timebuffer );
-        t->tv_sec = timebuffer.time;
+        t->tv_sec = (long)timebuffer.time;
         t->tv_usec = 1000 * timebuffer.millitm;
     }
 #else
